+++ vm/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2026-02-27
+
 ### Added
 
 - Add `sha256` and `verify_kzg_proof` host queries [#3993]
@@ -96,7 +98,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3405]: https://github.com/dusk-network/rusk/issues/3405
 [#3437]: https://github.com/dusk-network/rusk/issues/3437
 
-[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.3...HEAD
+[Unreleased]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.6.0...HEAD
+[1.6.0]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.3...dusk-vm-1.6.0
 [1.4.3]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.2...dusk-vm-1.4.3
 [1.4.2]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.1...dusk-vm-1.4.2
 [1.4.1]: https://github.com/dusk-network/rusk/compare/dusk-vm-1.4.0...dusk-vm-1.4.1